% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODA.R
\docType{data}
\name{datExpr1}
\alias{datExpr1}
\title{datExpr1}
\format{A matrix with 20 rows and 500 columns.}
\description{
Synthetic gene expression profile with 20 samples and 500 genes.
}
\examples{
data(synthetic)
## plot the heatmap of the correlation matrix ...
\dontrun{heatmap(cor(as.matrix(datExpr1)))}
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\keyword{data}

