\name{evppi}
\alias{evppi}
\title{Expected Value of Partial Pefect Information}
\usage{
evppi(x, G)
}
\arguments{
  \item{x}{the parameter simulation vector}

  \item{G}{the objective simuation matrix for all
  alternatives}
}
\description{
Function for efficiently computing the expected value of
partial perfect information (EVPPI) using Sadatssfavi et
al's method (see reference).
}
\seealso{
\code{\link{evppi_and_risk}}, \code{\link{evtpi}},
\code{\link{evtpi_and_risk}}
}

