% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getPartition}
\alias{getPartition}
\title{Get numeric partition from folder}
\usage{
getPartition(ResultFolder)
}
\arguments{
\item{ResultFolder}{folder used to save modules}
}
\value{
Number of partitions
}
\description{
Get identified partitionAssignment, only for synthetic data where gene names are numbers
}

