\name{linear_utility}
\alias{linear_utility}
\title{Linear Utility}
\usage{
linear_utility(value, zeroUtility, oneUtility)
}
\arguments{
  \item{value}{the value, or a vector of simulated values,
  associated with a goal}

  \item{zeroUtility}{the goal level with a utiltiy of zero}

  \item{oneUtility}{the goal level with a utility of one}
}
\description{
Returns the utility value associated to a goal level
assuming a linear utiltiy function for levels associated to
zero and one utilt.y
}

