% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODA.R
\name{MIcondition}
\alias{MIcondition}
\title{Modules detection by each condition}
\usage{
MIcondition(datExpr, conditionNames, ResultFolder, GeneNames, maxsize = 100,
  minsize = 30)
}
\arguments{
\item{datExpr}{gene expression profile, rows are samples and columns genes, 
rowname should contain condition specifier}

\item{conditionNames}{character vector, each as the condition name}

\item{ResultFolder}{where to store the clusters}

\item{GeneNames}{normally the gene official names to replace the colnames of datExpr}
}
\value{
a numeric vector, each entry is the number of modules in condition-specific network
}
\description{
Module detection on each condition-specific network, which is constructed from
all samples but samples belonging to that condition
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\keyword{multiplecondition}

