% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODA.R
\name{ModuleFrequency}
\alias{ModuleFrequency}
\title{Statistics of all conditions}
\usage{
ModuleFrequency(ResultFolder, intModules, conditionNames, indicator)
}
\arguments{
\item{ResultFolder}{where to store results}

\item{intModules}{how many modules in the background network}

\item{conditionNames}{a character vector, each of them is the name 
of condition. Or just single name}

\item{indicator}{identifier of current profile, served as a tag in name}
}
\value{
None
}
\description{
Statistics of all conditions. To highlight conserved or condition-specific 
by counting how frequent each module is lablelled as which, and then 
visualize the frequency by bar plot.
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\seealso{
\code{\link{WeightedModulePartitionDensity}},
\code{\link{CompareAllNets}}
}
\keyword{Statistics}
\keyword{differential}
\keyword{module}

