% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODA.R
\name{modulesRank}
\alias{modulesRank}
\title{Modules rank}
\usage{
modulesRank(W, modulefile, GeneNames)
}
\arguments{
\item{W}{Edges weights matrix for WGCN}

\item{modulefile}{plain text, the same as savefile in \code{\link{recursiveigraph}}}

\item{GeneNames}{Gene symbols, sometimes we need them instead of probe ids}
}
\description{
Assign the module scores by weights, and rank them from highest to lowest
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\seealso{
\code{\link{recursiveigraph}}
}

