% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oda.R
\name{oda.lm}
\alias{oda.lm}
\title{Title}
\usage{
oda.lm(formula, data, subset, weights, na.action = "na.omit",
  n.models = NULL, betaprior = GDP(), modelprior = beta.binomial(1, 1),
  method = "coordinate-ascent", force.in, iterations = NULL)
}
\arguments{
\item{formula}{an object of class \code{\link{formula}} (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{weights}{an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector. If non-NULL, weighted least squares is used with weights weights (that is, minimizing sum(w*e^2)); otherwise ordinary least squares is used.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another possible value is NULL, no action. Value na.exclude can be useful.}

\item{n.models}{number of models to sample}

\item{betaprior}{prior on coefficients}

\item{modelprior}{prior on the model space}

\item{method}{the method to be used}

\item{force.in}{variables that shoujld always be included}

\item{iterations}{number of iterations}
}
\value{
None yet
}
\description{
Title
}
\details{
add more here
}
\examples{
library(MASS)
data(UScrime)
oda.lm(y ~ ., data=UScrime)

}
\references{
Clyde, MA, & Ghosh, J. (2012). Finite population estimators in stochastic search variable selection. Biometrika, 99 (4), 981-988. <10.1093/biomet/ass040>
}

