% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{recursiveigraph}
\alias{recursiveigraph}
\title{Modules identification by recursive community detection}
\usage{
recursiveigraph(g, savefile, method = c("fastgreedy", "louvain"),
  maxsize = 200, minsize = 30)
}
\arguments{
\item{g}{igraph object, the network to be partitioned}

\item{savefile}{plain text, used to store module, each line as a module}

\item{method}{specify the community detection algorithm}

\item{maxsize}{maximal module size}

\item{minsize}{minimal module size}
}
\description{
Modules detection using igraph's community detection algorithms, when the
resulted module is larger than expected, it is further devided by the same program
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\references{
Blondel, Vincent D., et al. "Fast unfolding of communities in 
large networks." Journal of statistical mechanics: theory and experiment 
2008.10 (2008): P10008.
}
\keyword{community}
\keyword{detection}

