\name{relative2absolute}
\alias{absolute2relative}
\alias{relative2absolute}
\title{Converts between relative and absolute decision vectors}
\usage{
relative2absolute(A, DSpec)

absolute2relative(A, DSpec)
}
\arguments{
  \item{A}{one or more decision vectors (relative or
  absolute)}

  \item{DSpec}{a vector specifying the number of options in
  each decision}
}
\description{
Converts between relative and absolute decision vectors
}

