\name{relative2named}
\alias{relative2named}
\title{Transform a relativde decision vector into a named decision vector}
\usage{
relative2named(A, DList)
}
\arguments{
  \item{A}{one or more relative decision vectors}

  \item{DList}{a list of decsions where each decision has a
  vector of possible options}
}
\description{
Transform a relativde decision vector into a named decision
vector
}

