\name{sd_from_90CI}
\alias{sd_from_90CI}
\title{Standard Deviation of a 90\% Confidence Interval}
\usage{
sd_from_90CI(lower, upper)
}
\arguments{
  \item{lower}{the lower bound of the 90\% confidence
  interval}

  \item{upper}{the upper bound of the 90\% confidence
  interval}
}
\description{
Generates the standard deviation from the lower and upper
bounds of a 90\% confidence interval
}

