\name{sumproduct}
\alias{sumproduct}
\title{Sum vroduct of vectors and matrices}
\usage{
sumproduct(x, y)
}
\arguments{
  \item{x}{a vector or matrix}

  \item{y}{a vector of matrix whose dimensions are
  compatible with x for a sum product operation}
}
\description{
Vectorized implementation of the sumproduct of vector and
matrices
}

