% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODA.R
\name{WeightedModuleLouvain}
\alias{WeightedModuleLouvain}
\title{Modules detection by Louvain algorithm}
\usage{
WeightedModuleLouvain(datExpr, foldername, indicatename, GeneNames,
  maxsize = 100, minsize = 3, power = 10)
}
\arguments{
\item{datExpr}{gene expression profile, rows are samples and columns genes}

\item{foldername}{where to store the clusters}

\item{indicatename}{normally a specific tag of condition}

\item{maxsize}{the maximal nodes allowed in one module}

\item{minsize}{the minimal nodes allowed in one module}

\item{power}{the power parameter of WGCNA, W_{ij}=|cor(x_i,x_j)|^power}
}
\value{
The number of clusters
}
\description{
Module detection based on the Louvain algorithm, which tries to maximize 
overall modularity of resulting partition.
}
\examples{
data(synthetic)
ResultFolder = 'ForSynthetic' # where middle files are stored
CuttingCriterion = 'Density' # could be Density or Modularity
indicator1 = 'X'     # indicator for data profile 1
indicator2 = 'Y'      # indicator for data profile 2
specificTheta = 0.1 #threshold to define condition specific modules
conservedTheta = 0.1#threshold to define conserved modules
intModules1 <- WeightedModulePartitionDensity(datExpr1,ResultFolder,
indicator1,CuttingCriterion) 
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\references{
Blondel, Vincent D., et al. "Fast unfolding of communities in 
large networks." Journal of statistical mechanics: theory and experiment 
2008.10 (2008): P10008.
}
\keyword{cutting}
\keyword{dendrogram}

