% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{modulesRank}
\alias{modulesRank}
\title{Modules rank from recursive communities detection}
\usage{
modulesRank(foldername, indicator, GeneNames)
}
\arguments{
\item{foldername}{folder used to save modules}

\item{indicator}{normally a specific tag of condition}

\item{GeneNames}{Gene symbols, sometimes we need them instead of probe ids}
}
\value{
The numeber of modules
}
\description{
Assign the module scores by weights, and rank them from highest to lowest
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\seealso{
\code{\link{recursiveigraph}}
}

